requirements.txt
numpy==1.19.5
pandas==1.3.5
Pillow
PyYAML==6.0
scipy==1.7.3
torch==1.10.1
torchtext==0.11.1
torchvision==0.11.2
transformers==4.15.0
wandb
tensorboard
gpustat
matplotlib
scipy
seaborn
tqdm
sklearn
sentencepiece
avalanche-lib
datasets
setuptools==59.5.0